refactor: script refactoring with implement of es6-modules
