+++ src/JS/key.js
@@ -1,0 +1,9 @@
+class Key {
+    constructor(name, isFunctional, code) {
+        this.name = name;
+        this.isFunctional = isFunctional;
+        this.code = code;
+    }
+}
+
+export { Key };
